Creating a fix for sssd's lib
See https://github.com/Linuxbrew/legacy-linuxbrew/issues/500 for more information. Basically the linuxbrew installed glibc cannot communicate with ldap based users for accessing things over getpwent. The right solution is to build this. The quick and dirty solution is we accept this happens and pull in a symbolic link to libnss_sss.so.2 into the homebrew glibc lib.