Bring in readline and libtool dependency.
When using libedit the CMake dependency toolchain had an issue with the test for EDITLINE_HAVE_COMPLETION. libedit lacks the right routine for rl_completion_entry_function but readline has it. Switching to using readline.

Also we noticed libtool was a missing dependency even though it was explicitly used in the install subroutine.