Switch to using bundled libedit versus system.
We tried to use the libedit that ships with homebrew/dupes/libedit and get MySQL to path to it. MySQL's cmake did find the linuxbrew libedit but the cmake tests failed in EDITLINE_HAVE_COMPLETION because we could not find the rl_completion_entry_function somehow. This is in libedit's `include/editline/readline.h` and the string can be found in the libedit .so and .a files. But mysql will not have it. So switching to bundled and the library and client compiles.