License and tag changes.
As part of the migration we are going to move code, which has come from another brew repo out into its own repo. That code will be licensed under the terms of homebrew whilst this repo will retain Ensembl licensing. All files editted in this change are not subject to the move.